Move most of Private.hs to Server.hs.
This is similar to what was done to Public.hs, to re-organize the
routes, and in particular removes the "public/" or "private/" prefixes.
Things can be properly namespaced later.
